[FIXED HUDSON-6351] update DirectoryBrowserSupport to use long for contentLength, and update stapler to 1.141 which supports long for contentLength.
git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/main@30571 71c3de6d-444a-0410-be80-ed276b4c234a

+++ core/src/main/java/hudson/model/DirectoryBrowserSupport.java
@@ -288,11 +288,11 @@ public final class DirectoryBrowserSupport implements HttpResponse {
     }
 
     private static final class ContentInfo implements FileCallable<ContentInfo> {
-        int contentLength;
+        long contentLength;
         long lastModified;
 
         public ContentInfo invoke(File f, VirtualChannel channel) throws IOException {
-            contentLength = (int) f.length();
+            contentLength = f.length();
             lastModified = f.lastModified();
             return this;
         }
